requirements.txt
# PDF extraction
pymupdf>=1.23.0

# Vector database
chromadb>=0.4.24

# Embeddings
sentence-transformers>=2.7.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# LLM
anthropic>=0.26.0

# Config
python-dotenv>=1.0.0